Allow alternative protocol specification
When Git encounters a URL of the form <transport>::<address>, Git
invokes git-remote-<transport> with <address> as the second argument.

However, when Git encounters a URL of the form <transport>://<address>,
Git invokes git-remote-<transport> with the full URL as the second
argument.

Previously, we only supported the first form. This patch adds support
for the <transport>://<address> form so that either
dropbox::/path/to/repository or dropbox://path/to/repository may be
used.