Fix Windows compatibility
This patch fixes two behaviors that were broken on Windows.

When dealing with remote paths, we used `os.path.join` to join path
elements. The behavior of this function is dependent on the OS: it joins
elements using `os.path.sep`, which is '/' on POSIX systems but '\' on
Windows.

When constructing remote paths, we always want to use '/' as a path
separator, regardless of the setting of `os.path.sep` on the local
system. This patch fixes this inconsistency by replacing `os.path.join`
with `posixpath.join`.

The Git remote helper communication protocol is strict in how it deals
with newlines: it expects '\n' newlines, and it does not handle '\r\n'
newlines gracefully. Any '\r' characters at the end of a line are
treated as part of the contents of that line. Python opens stdout in
text mode, and on Windows, '\n' line endings are automatically converted
to '\r\n' line endings, breaking communication.

This patch sets the line-end translation mode for stdout to binary on
Windows so that '\n' characters are left as-is. Note that we do not have
to change the mode for stderr: it is left in text mode because it is
connected to the terminal for error reporting.