Make helper terminate worker processes properly
Previously, when there were errors while downloading, the helper would
print out an error message and then hang, requiring the user to manually
terminate the process. This was caused by worker processes that wouldn't
automatically exit when the main process exited.

This patch fixes the behavior by setting the `daemon` flag on child
processes so that when the main process exits, it attempts to terminate
all of its worker processes.