Fix bug with initializing new repositories remotely
For some reason the "error" field was missing from an error returned by
the Dropbox API. This was causing the initial push to fail with an
"error: unexpected exception".

This is a hack that fixed it for me. This is probably not the best way
but it's something. Could be worth looking into the source of this
problem more.