Fix stdout on Windows
When git-remote-dropbox is installed using `setup.py`, the entry point
becomes `main()` directly. This makes it so that the setup code under
`if __name__ == '__main__':` doesn't execute.

This change moves the code that converts stdout to binary into `main()`,
which means that it will always be called regardless of how the module
is executed.