Fix no default remote head issue
When cloning a repository it is expected to have a default branch,
that is checked out after cloning. That default branch is stored
as a symbolic ref in `path-to-repo/HEAD`.
This change the `do_list` to list that symbolic ref as well.

When the first push occurs to a dropbox remote, this change create
the symbolic ref to point at the current local branch's remote pair
or if that is not pushed, then the first pushed branch.