Fix push error message description
When the remote helper detects error while pushing, it should state
that in a `error <dst> <why>` manner.
The client display more descriptive error if the `why` is in some
predefined set of strings, such as `fetch first`, `non-fast forward`
(see `transport-helper.c` in the git client's source).

The remote helper incorrectly reports `non-fast-forward` instead of
`non-fast forward`.
When the remote branch's tip is not present locally it should be report
that with `fetch first`.

This commit fix thoose issues.