Fix handling of remote HEAD
This patch is derived from the contributions of Dávid Kis
<cf.paintech@gmail.com> and Niclas Eriksson <niclas.eriksson@knowit.se>.

This patch maintains a default branch on the remote, so that the correct
branch is automatically checked out upon cloning. It also prevents
deletion of the remote branch.

This patch does not provide a mechanism for updating the default branch:
it must be handled out-of-band by manually editing the `HEAD` file in
Dropbox.