Fix bug related to empty tree
When using git-remote-dropbox with versions of Git prior to
f06ab027efd2cff4c4319354f1ad7d5f09e853a1, if a repository being cloned
contained the empty tree, then git-remote-dropbox would crash while
cloning. This was reported in
https://github.com/anishathalye/git-remote-dropbox/issues/82.

This happened because `git.object_exists()` returns true for the hash of
the empty tree (even when it is not present in the object store), but
`git.history_exists()` used to return false (prior to f06ab027), so
git-remote-dropbox would try to complete the history. This would lead to
a call of `git.referenced_objects()`, which due to a bug in
git-remote-dropbox, would crash when called with the empty tree. This
bug in git-remote-dropbox is fixed in this patch.

According to the commit message in
f06ab027efd2cff4c4319354f1ad7d5f09e853a1 in Git, it seems that Git tries
to support the workflow when the empty tree object is not present in the
object store even if it's referenced in the commit graph, even though
this is not recommended. However, `git fsck` will give a warning in this
situation. In order to avoid the warning, we explicitly add the empty
tree to the object store if it is referenced.

This patch also makes a minor change to the way `git cat-file` is called
when checking to see if an object exists: there is a '-e' switch
designed for checking if an object exists and is a valid object.