Enable more strict conflict checking
Thanks to an anonymous comment on Lobste.rs for this recommendation [1].
The `strict_conflict` parameter makes Dropbox "more strict about how
each WriteMode detects conflict" [2]. It doesn't pin down exactly what
this means, and the two examples it gives for non-strict conflict
checking are not an issue for git-remote-dropbox (a write/delete
conflict is not an issue, and a write/write conflict with identical
contents is not an issue). Still, this seems nice to enable, just in
case.

[1]: https://lobste.rs/s/5qqcev/git_remote_dropbox_transparent_bridge#c_tupdl1
[2]: https://www.dropbox.com/developers/documentation/http/documentation#files-upload